spectrogram generates erronous data but something
